+++ packages/medicines/requests/variant_requests.go
@@ -7,8 +7,8 @@ import (
 type SearchVariantRequest struct {
 	Name string `json:"Name" example:"name"`
 
-	Limit  int64      `json:"limit" example:"10"`
-	Offset int64      `json:"offset" example:"0"`
+	Limit  int        `json:"limit" example:"10"`
+	Offset int        `json:"offset" example:"0"`
 	Sort   SortOption `json:"sort"`
 }
